+++ ipv6logconv/ipv6logconv.c
@@ -2,7 +2,7 @@
  * Project    : ipv6calc
  * File       : ipv6logconv.c
  * Version    : $Id$
- * Copyright  : 2002-2015 by Peter Bieringer <pb (at) bieringer.de>
+ * Copyright  : 2002-2016 by Peter Bieringer <pb (at) bieringer.de>
  * 
  * Information:
  *  Dedicated program for logfile conversions
@@ -467,7 +467,7 @@ static int converttoken(char *resultstring, const size_t resultstring_length, co
 	DEBUGSECTION_BEGIN(DEBUG_ipv6logconv_processing)
 		if (inputtype >= 0) {
 			for (i = 0; i < MAXENTRIES_ARRAY(ipv6calc_formatstrings); i++) {
-				if (inputtype == ipv6calc_formatstrings[i].number) {
+				if ((uint32_t) inputtype == ipv6calc_formatstrings[i].number) {
 					DEBUGPRINT_WA(DEBUG_ipv6logconv_processing, "Found type: %s", ipv6calc_formatstrings[i].token);
 				};
 				break;
